Use a shared base config for local and remote webdriver tests.
